# packages/api-client/openapi.json
{
  "components" : {
    "schemas" : {
      "KeyTerm" : {
        "properties" : {
          "meaning" : {
            "type" : "string"
          },
          "term" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SectionRef" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SectionView" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "storyCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "SiteView" : {
        "properties" : {
          "contactEmail" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SourceRef" : {
        "properties" : {
          "homepage" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SourceView" : {
        "properties" : {
          "homepage" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "StoryPage" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/StoryView"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "StoryView" : {
        "properties" : {
          "attribution" : {
            "type" : "string"
          },
          "dateEstimated" : {
            "type" : "boolean"
          },
          "headline" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "keyTerms" : {
            "items" : {
              "$ref" : "#/components/schemas/KeyTerm"
            },
            "type" : "array"
          },
          "originalUrl" : {
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "section" : {
            "$ref" : "#/components/schemas/SectionRef"
          },
          "source" : {
            "$ref" : "#/components/schemas/SourceRef"
          },
          "summary" : {
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/sections" : {
      "get" : {
        "operationId" : "sections",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SectionView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/v1/site" : {
      "get" : {
        "operationId" : "site",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SiteView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/v1/sources" : {
      "get" : {
        "operationId" : "sources",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SourceView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/v1/stories" : {
      "get" : {
        "operationId" : "stories",
        "parameters" : [ {
          "in" : "query",
          "name" : "section",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoryPage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    },
    "/api/v1/stories/{id}" : {
      "get" : {
        "operationId" : "story",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoryView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "catalog-controller" ]
      }
    }
  }
}